move 'config' package into 'apis' package
