+++ pkg/apis/pingcap/v1alpha1/tidb_config_wraper.go
@@ -18,7 +18,7 @@ import (
 	"reflect"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/tidb-operator/pkg/util/config"
+	"github.com/pingcap/tidb-operator/pkg/apis/util/config"
 	"github.com/pingcap/tidb-operator/pkg/util/toml"
 	"k8s.io/apimachinery/pkg/util/json"
 )
